move files, shorten themes, add few shot examples
